+++ libs/oidc/common/src/index.ts
@@ -3,3 +3,4 @@ export * from './lib/oidc-common';
 export * from './lib/modules/auth/auth.module';
 export * from './lib/strategies/jwt/jwt.strategy';
 export * from './lib/guards/jwt/jwt.guard';
+export * from './lib/guards/admin/admin.guard';
