+++ libs/oidc/common/src/lib/guards/admin/admin.guard.spec.ts
@@ -1,0 +1,7 @@
+import { AdminGuard } from './admin.guard';
+
+describe('AdminGuard', () => {
+  it('should be defined', () => {
+    expect(new AdminGuard()).toBeDefined();
+  });
+});
